refactor(integrations): return mount contract to evals; harness keeps utilities
The AgentMount/AgentRunToolSpec contract abstracts over evals' tool
surface registry — an evals concern, not an integrations one. It moves
back to evals/core/contracts/tool.ts verbatim. The shared harness
module keeps only what thin SDK adapters genuinely need: the merged
sanitizeErrorMessage, buildAllowlistedEnv, HarnessLogger, and
HarnessAdapterError. Also ignore evals run artifacts (.trajectories,
rubric cache) — generated output that was dirtying trees and fmt scans.
