+++ packages/evals/.gitignore
@@ -1,1 +1,4 @@
 .browserbase/
+.trajectories/
+.rubric-cache/
+packages/
