+++ packages/evals/core/contracts/tool.ts
@@ -1,5 +1,4 @@
 import type { ProbeEvidence } from "stagehand-v3";
-import type { AgentMount } from "@browserbasehq/stagehand-integrations/harness";
 import type { EvalLogger } from "../../logger.js";
 import type { ActionTarget, FocusedTarget, TargetKind, WaitSpec } from "./targets.js";
 import type { PageRepresentation, RepresentationOpts } from "./representation.js";
@@ -169,9 +168,52 @@ export interface CoreTool {
   start(input: ToolStartInput): Promise<ToolStartResult>;
 }
 
-export {
-  AGENT_RUN_TOOL_NAME,
-  AGENT_RUN_TOOL_RESERVED_HANDLES,
-  AGENT_RUN_TOOL_SERVER,
-} from "@browserbasehq/stagehand-integrations/harness";
-export type { AgentMount, AgentRunToolSpec } from "@browserbasehq/stagehand-integrations/harness";
+/**
+ * The MCP server / tool name used when an agent harness wraps handles in a
+ * code-execution tool.
+ */
+export const AGENT_RUN_TOOL_SERVER = "stagehand_browser";
+export const AGENT_RUN_TOOL_NAME = `mcp__${AGENT_RUN_TOOL_SERVER}__run`;
+export const AGENT_RUN_TOOL_RESERVED_HANDLES = ["startUrl", "task", "console"] as const;
+
+/**
+ * Surface-specific copy for the harness's code-execution tool. The harness
+ * owns mechanics and task bindings; the surface owns what the agent sees.
+ */
+export interface AgentRunToolSpec {
+  /** MCP tool description shown to the model. */
+  description: string;
+  /** Description of the tool's `code` parameter. */
+  codeParamDescription: string;
+  /** Message from the harness-owned tool allowlist when access is denied. */
+  denyMessage: string;
+}
+
+/**
+ * How an agent harness reaches an already-running surface. This is independent
+ * of `CoreTool.surface`; harnesses switch on `via` and need no surface-specific
+ * mounting logic.
+ */
+export type AgentMount = { promptInstructions: string } & (
+  | {
+      via: "handles";
+      /**
+       * Named values placed in snippet scope. Names, not order, bind values.
+       * `AGENT_RUN_TOOL_RESERVED_HANDLES` are injected by the harness and may
+       * not appear here.
+       */
+      handles: Record<string, unknown>;
+      runTool: AgentRunToolSpec;
+    }
+  | { via: "mcp"; mcpServers: Record<string, unknown> }
+  | {
+      via: "cli";
+      command: {
+        bin: string;
+        args?: string[];
+        cwd?: string;
+        /** Extra variables merged over the harness environment. */
+        env?: Record<string, string>;
+      };
+    }
+);
